refactor: add pagination component storybook
