+++ src/lib/hooks/common/usePage.ts
@@ -1,0 +1,8 @@
+import { useRouter } from "next/router";
+
+export default function usePage() {
+  const router = useRouter();
+  const pageParam = Number(router.query?.page);
+  const currentPage = Number.isNaN(pageParam) ? 1 : pageParam;
+  return currentPage;
+}
