+++ src/stories/common/Pagination/usePagination.ts
@@ -1,14 +1,7 @@
-import { useRouter } from "next/router";
-import { useEffect, useState } from "react";
-import navigator from "@lib/modules/navigator";
 import { removeQuery, replaceQuery } from "@lib/modules/queryString";
-
-export const usePage = () => {
-  const router = useRouter();
-  const pageParam = Number(router.query?.page);
-  const currentPage = Number.isNaN(pageParam) ? 1 : pageParam;
-  return currentPage;
-};
+import navigator from "@lib/modules/navigator";
+import { useRouter } from "next/router";
+import { useState, useEffect } from "react";
 
 export default function usePagination({
   totalPages,
@@ -18,8 +11,9 @@ export default function usePagination({
   buttonNum: number;
 }) {
   const [offset, setOffset] = useState(1);
-  const currentPage = usePage();
   const router = useRouter();
+  const pageParam = Number(router.query?.page);
+  const currentPage = Number.isNaN(pageParam) ? 1 : pageParam;
   const changeCurrentPage = (page: number) => {
     const url =
       page === 1
