+++ src/stories/common/icons/ChevronRight.tsx
